Create new function to render elements
